Use float precision by default
This commit should always be applied on top of master in a branched called  default-to-float. This
makes it easier to use srl_projection in supereight.

+++ include/srl/typedefs.hpp
@@ -46,7 +46,7 @@
 
 namespace srl {
   /// Set the precision of all operations (float or double).
-  typedef double float_t;
+  typedef float float_t;
 
   template <int _Rows, int _Cols>
   using Matrixf = Eigen::Matrix<float_t, _Rows, _Cols>;
